Version 0.4
- Added a generic aggregate command
- Oram2016 Finished
- Sharp2019 under development
- Testing VCF input
