+++ oram2016.weights.multilocus.txt
@@ -1,9 +1,9 @@
 ODDSRATIO	ID_1	CHROM_1	POS_1	GENOTYPE_1	ID_2	CHROM_2	POS_2	GENOTYPE_2
-7.03	6:32605884	6	32605884	C:C	6:32681483	6	32681483	T:C
-21.98	6:32605884	6	32605884	C:C	6:32681483	6	32681483	C:C
-4.53	6:32605884	6	32605884	C:T	6:32681483	6	32681483	T:T
-48.18	6:32605884	6	32605884	C:T	6:32681483	6	32681483	T:C
-48.18	6:32605884	6	32605884	C:T	6:32681483	6	32681483	C:C
-21.12	6:32605884	6	32605884	T:T	6:32681483	6	32681483	T:T
-48.18	6:32605884	6	32605884	T:T	6:32681483	6	32681483	T:C
-48.18	6:32605884	6	32605884	T:T	6:32681483	6	32681483	C:C
+7.03	6:32605884_0	6	32605884	CC	6:32681483_1	6	32681483	TC
+21.98	6:32605884_0	6	32605884	CC	6:32681483_2	6	32681483	CC
+4.53	6:32605884_1	6	32605884	CT	6:32681483_0	6	32681483	TT
+48.18	6:32605884_1	6	32605884	CT	6:32681483_1	6	32681483	TC
+48.18	6:32605884_1	6	32605884	CT	6:32681483_2	6	32681483	CC
+21.12	6:32605884_2	6	32605884	TT	6:32681483_0	6	32681483	TT
+48.18	6:32605884_2	6	32605884	TT	6:32681483_1	6	32681483	TC
+48.18	6:32605884_2	6	32605884	TT	6:32681483_2	6	32681483	CC
